Flattened scoped services one level, moved different resources in separate files
