+++ ScoutSuite/providers/aws/configs/services.py
@@ -7,7 +7,7 @@ from ScoutSuite.providers.aws.services.cloudformation import CloudFormationConfi
 from ScoutSuite.providers.aws.services.cloudtrail import CloudTrailConfig
 from ScoutSuite.providers.aws.services.cloudwatch import CloudWatchConfig
 from ScoutSuite.providers.aws.services.directconnect import DirectConnectConfig
-from ScoutSuite.providers.aws.services.ec2 import EC2
+from ScoutSuite.providers.aws.services.ec2.service import EC2
 from ScoutSuite.providers.aws.services.efs import EFSConfig
 from ScoutSuite.providers.aws.services.elasticache import ElastiCacheConfig
 from ScoutSuite.providers.aws.services.elb import ELBConfig
